add support for using existing VPC for ipv6
